# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_create_job_api_jobs_post": {
        "properties": {
          "extras": {
            "default": "",
            "title": "Extras",
            "type": "string"
          },
          "frame_limit": {
            "default": 6,
            "title": "Frame Limit",
            "type": "integer"
          },
          "local_whisper_compute_type": {
            "default": "",
            "title": "Local Whisper Compute Type",
            "type": "string"
          },
          "local_whisper_device": {
            "default": "",
            "title": "Local Whisper Device",
            "type": "string"
          },
          "note_base_url": {
            "default": "https://api.openai.com/v1",
            "title": "Note Base Url",
            "type": "string"
          },
          "note_language": {
            "$ref": "#/components/schemas/NoteLanguage"
          },
          "note_model": {
            "default": "gpt-5.5",
            "title": "Note Model",
            "type": "string"
          },
          "note_style": {
            "$ref": "#/components/schemas/NoteStyle",
            "default": "detailed"
          },
          "performance_mode": {
            "$ref": "#/components/schemas/PerformanceMode",
            "default": "balanced"
          },
          "subtitle": {
            "anyOf": [
              {
                "format": "binary",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "transcription_api_key": {
            "default": "",
            "title": "Transcription Api Key",
            "type": "string"
          },
          "transcription_base_url": {
            "default": "https://api.openai.com/v1",
            "title": "Transcription Base Url",
            "type": "string"
          },
          "transcription_language": {
            "$ref": "#/components/schemas/TranscriptionLanguage",
            "default": "auto"
          },
          "transcription_mode": {
            "$ref": "#/components/schemas/TranscriptionMode",
            "default": "audio_transcriptions"
          },
          "transcription_model": {
            "default": "whisper-1",
            "title": "Transcription Model",
            "type": "string"
          },
          "video": {
            "format": "binary",
            "title": "Video",
            "type": "string"
          }
        },
        "required": [
          "video",
          "note_language"
        ],
        "title": "Body_create_job_api_jobs_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobHistory": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobSummary"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "title": "JobHistory",
        "type": "object"
      },
...
      "NoteLanguage": {
        "enum": [
          "zh",
          "en",
          "follow"
        ],
        "title": "NoteLanguage",
        "type": "string"
      },
      "NoteStyle": {
        "enum": [
          "minimal",
          "detailed",
          "tutorial",
          "academic",
          "task_oriented",
          "meeting_minutes"
        ],
        "title": "NoteStyle",
        "type": "string"
      },
...
      "PerformanceMode": {
        "enum": [
          "fast",
          "balanced",
          "accurate"
        ],
        "title": "PerformanceMode",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Video Note Generator MVP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobHistory"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job_api_jobs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_job_api_jobs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Job Api Jobs Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/frame-suggestion": {
      "post": {
        "operationId": "suggest_frame_count_api_jobs_frame_suggestion_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_suggest_frame_count_api_jobs_frame_suggestion_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameSuggestion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Suggest Frame Count",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "operationId": "delete_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Job Api Jobs  Job Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPublicState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "cancel_job_api_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobPublicState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/storage": {
      "get": {
        "operationId": "get_job_storage_api_jobs__job_id__storage_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Job Storage Api Jobs  Job Id  Storage Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Storage",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/jobs/{job_id}/transcription/cache": {
      "delete": {
        "operationId": "delete_transcription_cache_api_jobs__job_id__transcription_cache_delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Transcription Cache Api Jobs  Job Id  Transcription Cache Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Transcription Cache",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/transcription/resume": {
      "post": {
        "operationId": "resume_transcription_api_jobs__job_id__transcription_resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Resume Transcription Api Jobs  Job Id  Transcription Resume Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Transcription",
        "tags": [
          "jobs"
        ]
      }
    },
...
  }
}